refactored index & edit pages into common page and modified few things as per requests
